Link users with journal entries

+++ src/main/java/com/project/journalProject/service/JournalEntryService.java
@@ -1,11 +1,14 @@
 package com.project.journalProject.service;
 
+import com.project.journalProject.entity.UserEntry;
 import com.project.journalProject.repository.JournalEntryRepository;
 import com.project.journalProject.entity.JournalEntry;
+import com.project.journalProject.repository.UserEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,51 +16,72 @@ import java.util.Optional;
 public class JournalEntryService {
 
     private final JournalEntryRepository journalEntryRepository;
+    private final UserEntryRepository userEntryRepository;
 
     @Autowired
-    public JournalEntryService(JournalEntryRepository journalEntryRepository) {
+    public JournalEntryService(JournalEntryRepository journalEntryRepository, UserEntryRepository userEntryRepository) {
         this.journalEntryRepository = journalEntryRepository;
+        this.userEntryRepository = userEntryRepository;
     }
 
-    public JournalEntry createEntry(JournalEntry journalEntry) {
-        return journalEntryRepository.save(journalEntry);
+    public JournalEntry createEntry(JournalEntry journalEntry, String userName) {
+        Optional<UserEntry> userEntry = userEntryRepository.findByUserName(userName);
+
+        if (userEntry.isPresent()) {
+            journalEntry.setDate(LocalDateTime.now());
+
+            JournalEntry savedEntry = journalEntryRepository.save(journalEntry);
+            userEntry.get().getJournalEntryList().add(savedEntry);
+            userEntryRepository.save(userEntry.get());
+
+            return savedEntry;
+        }
+        return null;
     }
 
-    public List<JournalEntry> getAll() {
-        return journalEntryRepository.findAll();
+    public List<JournalEntry> getAllJournalEntriesForUser(String userName) {
+        Optional<UserEntry> userEntry = userEntryRepository.findByUserName(userName);
+        return userEntry.map(UserEntry::getJournalEntryList).orElse(null);
     }
 
-    public Optional<JournalEntry> getEntryByID(ObjectId id) {
-        return journalEntryRepository.findById(id);
-    }
+    public Optional<JournalEntry> deleteByUserNameAndID(String userName, ObjectId id) {
+        Optional<UserEntry> entry = userEntryRepository.findByUserName(userName);
 
-    public Optional<JournalEntry> deleteByID(ObjectId id) {
-        Optional<JournalEntry> entry = journalEntryRepository.findById(id);
+        if (entry.isPresent()) {
+            JournalEntry journalEntry = entry.get().getJournalEntryList().stream()
+                    .filter(x -> x.getId().equals(id)).findFirst().orElse(null);
 
-        if(entry.isPresent()) {
-            journalEntryRepository.deleteById(id);
-            return entry;
+            if (journalEntry != null) {
+                entry.get().getJournalEntryList().remove(journalEntry);
+                userEntryRepository.save(entry.get());
+                journalEntryRepository.deleteById(id);
+                return Optional.of(journalEntry);
+            }
         }
         return Optional.empty();
     }
 
-    public Optional<JournalEntry> updateEntryByID(ObjectId id, JournalEntry newEntry) {
-        Optional<JournalEntry> entry = journalEntryRepository.findById(id);
+    public Optional<JournalEntry> updateByNameAndID(String userName, ObjectId id, JournalEntry newEntry) {
+        Optional<UserEntry> entry = userEntryRepository.findByUserName(userName);
 
         if(entry.isPresent()) {
-            JournalEntry oldEntry = entry.get();
+            JournalEntry journalEntry = entry.get().getJournalEntryList().stream().
+                    filter(x -> x.getId().equals(id)).findFirst().orElse(null);
 
-            if (!newEntry.getTitle().isBlank()) {
-                oldEntry.setTitle(newEntry.getTitle());
+            if (journalEntry == null) {
+                return Optional.empty();
             }
-            if (!newEntry.getContent().isBlank()) {
-                oldEntry.setContent(newEntry.getContent());
+            if (!newEntry.getTitle().isBlank()) {
+                journalEntry.setTitle(newEntry.getTitle());
+            }
+            if (newEntry.getContent() != null && !newEntry.getContent().isBlank()) {
+                journalEntry.setContent(newEntry.getContent());
             }
             if (newEntry.getDate() != null) {
-                oldEntry.setDate(newEntry.getDate());
+                journalEntry.setDate(newEntry.getDate());
             }
 
-            return Optional.of(journalEntryRepository.save(oldEntry));
+            return Optional.of(journalEntryRepository.save(journalEntry));
         }
         return Optional.empty();
     }
